+++ kite-core/src/main/kotlin/com/tang/kite/enumeration/IndexOrder.kt
@@ -1,0 +1,14 @@
+package com.tang.kite.enumeration
+
+/**
+ * Index sort order
+ *
+ * @author Tang
+ */
+enum class IndexOrder {
+
+    ASC,
+
+    DESC;
+
+}