+++ kite-core/src/test/kotlin/com/tang/kite/schema/builder/AccountBuilder.kt
@@ -7,21 +7,21 @@ import com.tang.kite.annotation.field.CreateTime
 import com.tang.kite.annotation.field.UpdateTime
 import com.tang.kite.annotation.id.Id
 import com.tang.kite.annotation.id.IdType
-import com.tang.kite.enumeration.SortOrder
+import com.tang.kite.enumeration.IndexOrder
 import java.math.BigDecimal
 import java.time.LocalDateTime
 
 /**
  * @author Tang
  */
-@CompositeIndex(unique = true, columns = ["username", "nickname"], orders = [SortOrder.ASC, SortOrder.DESC])
+@CompositeIndex(unique = true, columns = ["username", "nickname"], orders = [IndexOrder.ASC, IndexOrder.DESC])
 @Table(value = "account", comment = "Table for account builder")
 class AccountBuilder(
 
     @Id(type = IdType.AUTO)
     var id: Long? = null,
 
-    @Index(order = SortOrder.DESC)
+    @Index(order = IndexOrder.DESC)
     var username: String? = null,
 
     var nickname: String? = null,