+++ kite-core/src/test/kotlin/com/tang/kite/schema/synchronization/index/addindex/AccountAddIndex.kt
@@ -7,7 +7,7 @@ import com.tang.kite.annotation.field.CreateTime
 import com.tang.kite.annotation.field.UpdateTime
 import com.tang.kite.annotation.id.Id
 import com.tang.kite.annotation.id.IdType
-import com.tang.kite.enumeration.SortOrder
+import com.tang.kite.enumeration.IndexOrder
 import java.time.LocalDateTime
 
 /**
@@ -15,7 +15,7 @@ import java.time.LocalDateTime
  *
  * @author Tang
  */
-@CompositeIndex(unique = true, columns = ["username", "nickname"], orders = [SortOrder.ASC, SortOrder.DESC])
+@CompositeIndex(unique = true, columns = ["username", "nickname"], orders = [IndexOrder.ASC, IndexOrder.DESC])
 @Table(value = "account", comment = "Table for account add index")
 class AccountAddIndex(